changed entity null to max-value elements

+++ SerenityLib/src/serenity/ecs/entity/Entity.cpp
@@ -48,7 +48,7 @@ std::vector<Entity> Entity::getChildren() const noexcept {
     if (transform) {
         auto& ecs = *Engine::priv::PublicEntity::GetECS(*this);
         auto& pcs = ecs.getSystem<SystemTransformParentChild>();
-        for (uint32_t i = 0; i < (uint32_t)pcs.m_Parents.size(); ++i) {
+        for (size_t i = 0; i < pcs.m_Parents.size(); ++i) {
             const auto parentID = pcs.m_Parents[i];
             if (parentID == m_ID) {
                 output.push_back(ecs.getEntityPool().getEntityFromID(i + 1));
